UI 디테일 수정 및 코드 정리
- UI 수정
- 텍스트 수정
- 코드 정리

+++ app/src/main/java/com/nexters/frutto/view/viewholder/ApplicantListViewHolder.kt
@@ -37,7 +37,7 @@ class ApplicantListViewHolder(
                     "취소",
                     "내보내기"
                 ) {
-                    onExitApplicant(applicant.room.id.toLong(), applicant.user.id)
+                    onExitApplicant(applicant.room.id, applicant.user.id)
                 }.show(supportFragmentManager, TAG)
             }
